# he/kms/openapi_client/models/pagination_metadata.py
# coding: utf-8

"""
    Key Manager API

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 1.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json

from pydantic import BaseModel, ConfigDict, StrictInt
from typing import Any, ClassVar, Dict, List, Optional
from typing import Optional, Set
from typing_extensions import Self

class PaginationMetadata(BaseModel):
    """
    PaginationMetadata
    """ # noqa: E501
    first_page: Optional[StrictInt] = None
    last_page: Optional[StrictInt] = None
    next_page: Optional[StrictInt] = None
    page: Optional[StrictInt] = None
    previous_page: Optional[StrictInt] = None
    total: Optional[StrictInt] = None
    total_pages: Optional[StrictInt] = None
    __properties: ClassVar[List[str]] = ["first_page", "last_page", "next_page", "page", "previous_page", "total", "total_pages"]

    model_config = ConfigDict(
        populate_by_name=True,
        validate_assignment=True,
        protected_namespaces=(),
    )


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Optional[Self]:
        """Create an instance of PaginationMetadata from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        excluded_fields: Set[str] = set([
        ])

        _dict = self.model_dump(
            by_alias=True,
            exclude=excluded_fields,
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Optional[Dict[str, Any]]) -> Optional[Self]:
        """Create an instance of PaginationMetadata from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "first_page": obj.get("first_page"),
            "last_page": obj.get("last_page"),
            "next_page": obj.get("next_page"),
            "page": obj.get("page"),
            "previous_page": obj.get("previous_page"),
            "total": obj.get("total"),
            "total_pages": obj.get("total_pages")
        })
        return _obj