# src/lib/kubb-he/types/GetMessages200.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { Message } from "./Message";
import type { Meta } from "./Meta";

export type GetMessages200 = {
  /**
   * @type array
   */
  array: Message[];
  /**
   * @type object
   */
  meta: Meta;
};